@@ -37,19 +37,22 @@
     "ascii-progress": "^1.0.5",
     "body-parser": "^1.18.2",
     "btoa": "^1.1.2",
-    "checksum": "^0.1.1",
     "cookie-parser": "^1.4.3",
-    "cytoscape": "git+https://github.com/d2fong/cytoscape.js.git#mount",
+    "express": "^4.16.2",
+    "express-jwt": "^5.3.0",
+    "classnames": "^2.2.5",
+    "cytoscape": "git+https://github.com/d2fong/cytoscape.js.git#unstable-build",
     "cytoscape-cola": "^2.0.0",
     "cytoscape-compound-collapse": "git+https://github.com/d2fong/cytoscape.js-compound-collapse.git",
     "cytoscape-cose-bilkent": "git+https://github.com/d2fong/cytoscape.js-cose-bilkent.git#simple-label-dimensions",
-    "cytoscape-dagre": "^2.1.0",
+    "cytoscape-dagre": "^2.0.0",
     "cytoscape-fisheye": "git+https://github.com/d2fong/cytoscape.js-fisheye.git",
     "cytoscape-klay": "^3.0.0",
     "cytoscape-sbgn-stylesheet": "^3.0.1",
-    "express": "^4.16.2",
-    "express-jwt": "^5.3.0",
+    "ejs": "^2.5.7",
+    "file-saver": "^1.3.3",
     "fready": "^1.0.0",
+    "gzip-js": "^0.3.2",
     "hyperscript": "^2.0.2",
     "ip": "^1.1.5",
     "json-hash": "^1.2.0",
@@ -60,13 +63,13 @@
     "node-fetch": "^1.7.3",
     "pathway-commons": "^1.5.4",
     "query-string": "^5.0.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-ga": "^2.3.4",
     "react-hyperscript": "^3.0.0",
     "react-router-dom": "^4.2.2",
     "rethinkdb": "^2.3.3",
-    "sbgnml-to-cytoscape": "^3.0.2",
+    "sbgnml-to-cytoscape": "^4.0.1",
     "serve-favicon": "^2.4.5",
     "socket.io": "^2.0.3",
     "tippy.js": "^1.3.0",
