@@ -18,12 +18,7 @@ class EditWarning extends React.Component {
     return this.props.active !== nextProps.active;
   }
 
-  componentWillUpdate() {
-    document.getElementsByClassName('editWarning')[0].style.display = (this.props.active ? 'flex' : 'none');
-  }
-
-  // Dynamically set "left" to centre div on screen after mounting
-  componentDidMount() {
+  componentDidUpdate() {
     const warningDOM = document.getElementsByClassName('editWarning')[0];
     
     // Set centering style for warning
