@@ -1,12 +1,9 @@
-const bindMove = (cy) => {
-  // Mobile finish dragging node event
-  // cy.on('tapdragout', 'node', function(evt) {
-  //   console.log('Stopped dragging the node:\n'+evt.target.id()+'\nAt position:\n'+JSON.stringify(evt.target.position()));
-  // });
+const CDC = require('../../../../services/index.js').CDC;
 
-  //
+const bindMove = (uri, version, editkey, cy) => {
+
   cy.on('free', 'node', function(evt) {
-    //console.log('Stopped dragging the node:\n'+evt.target.id()+'\nAt position:\n'+JSON.stringify(evt.target.position()));
+    console.log('Stopped dragging the node:\n'+evt.target.id()+'\nAt position:\n'+JSON.stringify(evt.target.position()));
   });
 
 };
