@@ -1,75 +1,9 @@
-module.exports = [
-  {
-    "id": "reactome",
-    "uri": "http://pathwaycommons.org/pc2/reactome",
-    "name": "Reactome",
-    "description": "Reactome v61 (only 'Homo_sapiens.owl') 23-Jun-2017",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/reactome.png"
-  },
-  {
-    "id": "pid",
-    "uri": "http://pathwaycommons.org/pc2/pid",
-    "name": "NCI Pathway Interaction Database: Pathway",
-    "description": "NCI Curated Human Pathways from PID (final); 27-Jul-2015",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/nci_nature.png"
-  },
-  {
-    "id": "humancyc",
-    "uri": "http://pathwaycommons.org/pc2/humancyc",
-    "name": "HumanCyc",
-    "description": "HumanCyc 20; 2016; under license from SRI International, www.biocyc.org",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/humancyc.png"
-  },
-  {
-    "id": "panther",
-    "uri": "http://pathwaycommons.org/pc2/panther",
-    "name": "PANTHER Pathway",
-    "description": "PANTHER Pathways 3.4.1 on 04-Jul-2016 (auto-converted to human-only model)",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/panther.png"
-  },
-  {
-    "id": "kegg",
-    "uri": "http://pathwaycommons.org/pc2/kegg",
-    "name": "KEGG Pathway",
-    "description": "KEGG 07/2011 (only human, hsa* files), converted to BioPAX by BioModels (http://www.ebi.ac.uk/biomodels-main/) team",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/kegg128.gif"
-  },
-  {
-    "id": "smpdb",
-    "uri": "http://pathwaycommons.org/pc2/smpdb",
-    "name": "Small Molecule Pathway Database",
-    "description": "Small Molecule Pathway Database 2.0, 05-Jun-2016",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/smpdb_icon.png"
-  },
-  {
-    "id": "inoh",
-    "uri": "http://pathwaycommons.org/pc2/inoh",
-    "name": "Integrating Network Objects with Hierarchies",
-    "description": "INOH 4.0 (signal transduction and metabolic data), 22-MAR-2011",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/inoh_logo.png"
-  },
-  {
-    "id": "netpath",
-    "uri": "http://pathwaycommons.org/pc2/netpath",
-    "name": "NetPath",
-    "description": "NetPath 12/2011",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/netpath_logo.png"
-  },
-  {
-    "id": "wp",
-    "uri": "http://pathwaycommons.org/pc2/wp",
-    "name": "WikiPathways",
-    "description": "WikiPathways - Community Curated Human Pathways; 29/09/2015 (human)",
-    "type": "BIOPAX",
-    "iconUrl": "http://pathwaycommons.github.io/cpath2/logos/wikipathways.png"
-  }
-];
+const pc = require('pathway-commons');
 
+const getPcDataSources = async () => {
+  const pcDataSources =  await pc.datasources.fetch();
+
+  return pcDataSources;
+}
+
+module.exports = getPcDataSources;
