@@ -39,16 +39,7 @@ const io = require('socket.io')(server);
 // view engine setup
 app.set('views', path.join(__dirname, '../', 'views'));
 
-// define an inexpensive html engine that doesn't do serverside templating
-app.engine('html', function (filePath, options, callback){
-  fs.readFile(filePath, function (err, content) {
-    if( err ){ return callback( err ); }
-
-    return callback( null, content.toString() );
-  });
-});
-
-app.set('view engine', 'html');
+app.set('view engine', 'ejs');
 
 app.use(favicon(path.join(__dirname, '../..', 'public', 'icon.png')));
 app.use(morgan('dev', {
@@ -95,7 +86,7 @@ app.use(function(req, res, next) {
 if (app.get('env') === 'development') {
   app.use(function(err, req, res, next) {
     res.status(err.status || 500);
-    res.render('error.html');
+    res.render('error');
   });
 }
 
@@ -104,7 +95,7 @@ if (app.get('env') === 'development') {
 // error page handler
 app.use(function(err, req, res, next) {
   res.status(err.status || 500);
-  res.render('error.html');
+  res.render('error');
 });
 
 
